refactor: move crates into rd-std
