--- rd-redir/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "rd-redir"
-version = "0.1.0"
-authors = ["spacemeowx2 <spacemeowx2@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-rd-interface = { version = "0.3", path = "../rd-interface" }
-futures = "0.3"
-async-net = "1.5"
-serde = "1.0"
-serde_derive = "1.0"
-libc = "0.2.91"
-socket2 = "0.4.0"
-log = "0.4"
-
-[features]
-default = []
-plugin = []
-
-[lib]
-crate-type = ["cdylib", "rlib"]
