+++ rd-std/src/redir/util.rs
