+++ rd-std/src/socks5/server.rs
@@ -1,12 +1,12 @@
 use super::common::{pack_udp, parse_udp, Address};
-use crate::protocol::{
+use super::protocol::{
     self, AuthMethod, AuthRequest, AuthResponse, CommandRequest, CommandResponse, Version,
 };
 use futures::{io::BufWriter, prelude::*};
 use protocol::Command;
 use rd_interface::{
-    async_trait, pool::IUdpChannel, util::connect_tcp, ConnectionPool, Context, IntoAddress,
-    IntoDyn, Net, Result, TcpStream, UdpSocket,
+    async_trait, pool::IUdpChannel, util::connect_tcp, ConnectionPool, Context, IServer,
+    IntoAddress, IntoDyn, Net, Result, TcpStream, UdpSocket,
 };
 use std::{
     net::{Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4},
@@ -176,5 +176,42 @@ impl IUdpChannel for Socks5UdpSocket {
         } else {
             0
         })
+    }
+}
+
+pub struct Socks5 {
+    server: Socks5Server,
+    listen_net: Net,
+    bind: String,
+}
+
+#[async_trait]
+impl IServer for Socks5 {
+    async fn start(&self, pool: ConnectionPool) -> Result<()> {
+        let listener = self
+            .listen_net
+            .tcp_bind(&mut Context::new(), self.bind.into_address()?)
+            .await?;
+
+        loop {
+            let (socket, addr) = listener.accept().await?;
+            let server = self.server.clone();
+            let pool2 = pool.clone();
+            let _ = pool.spawn(async move {
+                if let Err(e) = server.serve_connection(socket, addr, pool2).await {
+                    log::error!("Error when serve_connection: {:?}", e)
+                }
+            });
+        }
+    }
+}
+
+impl Socks5 {
+    pub fn new(listen_net: Net, net: Net, bind: String) -> Self {
+        Socks5 {
+            server: Socks5Server::new(listen_net.clone(), net),
+            listen_net,
+            bind,
+        }
     }
 }
